# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smaki CRM API Documentation",
        "version": "0.3"
    },
    "servers": [
        {
            "url": "https://api.smaki.com.ua/api"
        }
    ],
    "paths": {
...
        "/orders/{order_uuid}/courier": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Получить координаты курьера по заказку",
                "operationId": "59547859bcbf678f3be39f30a2c87da7",
                "parameters": [
                    {
                        "name": "order_uuid",
                        "in": "path",
                        "description": "UUID заказа в iiko",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "latitude": {
                                            "description": "Широта",
                                            "type": "string"
                                        },
                                        "longitude": {
                                            "description": "Долгота",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "data": {
                                            "latitude": "50.436518",
                                            "longitude": "30.5502122"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/products/{id}/main-image": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Получить главное фото для товара",
                "operationId": "9fd7db31029ede8c97314e4adc5a6c1e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID товара",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "product_id": {
                                            "description": "ID в формате UUID",
                                            "type": "string"
                                        },
                                        "file_name": {
                                            "description": "Полное название файла",
                                            "type": "string"
                                        },
                                        "name": {
                                            "description": "Название файла",
                                            "type": "string"
                                        },
                                        "mime_type": {
                                            "description": "Mime-Type",
                                            "type": "string"
                                        },
                                        "size": {
                                            "description": "Размер в байтах",
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "description": "Дата создания",
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "description": "Дата последнего редактирования",
                                            "type": "string"
                                        },
                                        "url": {
                                            "description": "Ссылка на фото",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "data": {
                                            "product_id": "5524e4ca-8655-41b8-9d79-b9fdbc7d39f4",
                                            "file_name": "gopher.png",
                                            "name": "gopher",
                                            "mime_type": "image/png",
                                            "size": "38456",
                                            "created_at": "2021-07-30T15:29:31.000000Z",
                                            "updated_at": "2021-07-30T15:29:31.000000Z",
                                            "url": "http://smaki.local/storage/29/goku.jpg"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Заменить главное фото для товара",
                "operationId": "b98733aeed23d31b4b1e40f43638b2e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID товара",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "Форматы: jpg, jpeg, png. Размер до 5МБ",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "product_id": {
                                            "description": "ID в формате UUID",
                                            "type": "string"
                                        },
                                        "file_name": {
                                            "description": "Полное название файла",
                                            "type": "string"
                                        },
                                        "name": {
                                            "description": "Название файла",
                                            "type": "string"
                                        },
                                        "mime_type": {
                                            "description": "Mime-Type",
                                            "type": "string"
                                        },
                                        "size": {
                                            "description": "Размер в байтах",
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "description": "Дата создания",
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "description": "Дата последнего редактирования",
                                            "type": "string"
                                        },
                                        "url": {
                                            "description": "Ссылка на фото",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "data": {
                                            "product_id": "5524e4ca-8655-41b8-9d79-b9fdbc7d39f4",
                                            "file_name": "gopher.png",
                                            "name": "gopher",
                                            "mime_type": "image/png",
                                            "size": "38456",
                                            "created_at": "2021-07-30T15:29:31.000000Z",
                                            "updated_at": "2021-07-30T15:29:31.000000Z",
                                            "url": "http://smaki.local/storage/29/goku.jpg"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}